Remove test to enable JS Crash Logging
Javascript crash logs could overwhelm the production server, and were
thus enabled only when connected to a QA server. The test relied on
the presence of a telephony related file, which will never exist for
current production targets, and may never exist again.

For now, the test has been commented out and marked with a TODO. The
implication is that JS crashes will not be logged.

No testing to perform as this was dead code.

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>

Change-Id: I6a5d6f2284c3aa195517af0a87e15ef990eacc95
Reviewed-on: https://g2g.palm.com/2663
Reviewed-by: Build Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Gayathri Srinivasan <gayathri.srinivasan@lge.com>