Supporting snapshot
:Release Notes:
Supporting snapshot for pmlogdaemon

:Detailed Notes:
pmlogdaemon is being started at the signal "init-boot-done".
It means can not record logs which are generated before "init-boot-done".
So, It needs to be moved into snapshot and now pmlogdaemon will cover
all of the logs.

:Testing Performed:
Compiled locally and tested on a device.
 - Checked pmlogdaemon is in snapshot.
 - Added script for recording some logs just after pmlogdaemon had started.
 - Checked the logs were recoreded well.

:QA Notes:

:Issues Addressed:
[BHV-11560] PmLogDaemon snapshot support

Open-webOS-DCO-1.0-Signed-off-by: Hongjeon Ha <hj.ha@lge.com>

Change-Id: I16ff7608cfa12ec5653f26cc18f458653fa5a52a
Reviewed-on: https://g2g.palm.com/6355
Reviewed-by: DCO Verification
Reviewed-by: Changhee Han <ch0.han@lge.com>
Reviewed-by: Hongjeon Ha <hj.ha@lge.com>
Tested-by: Hongjeon Ha <hj.ha@lge.com>
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Gayathri Srinivasan <gayathri.srinivasan@lge.com>